test: extract shared CDK test helpers, dedupe find-by-action logic
/simplify: extract findResourceEntryByLogicalIdPrefix/referencesLogicalId
into test/cdk-test-helpers.ts (was byte-identical duplicated across
akli-infrastructure.test.ts and app-site-stack.test.ts); generalize
findLambdaStatement into findStatementByAction and reuse it for the
new #227 invalidation assertions; hoist repeated policy lookups in
app-site-stack.test.ts's new describe block into a beforeEach.
